+++ library/Morloc/CodeGenerator/Grammars/Macro.hs
@@ -32,10 +32,10 @@ expandType
   -> (PVar -> [(PVar, MDoc)] -> MDoc) -- ^ make record type
   -> TypeP
   -> MDoc
-expandType mkfun mkrec t = f t where
+expandType mkfun mkrec t0 = f t0 where
   f :: TypeP -> MDoc
   f (VarP (PV _ _ v)) = pretty v
-  f t@(FunP t1 t2) = mkfun (f t1) (map f (typeArgs t))
+  f t@(FunP t1 _) = mkfun (f t1) (map f (typeArgs t))
   f (ArrP (PV _ _ v) ts) = pretty $ expandMacro v (map (render . f) ts)
   f (NamP _ v _ entries) = mkrec v [(k, f t) | (k, t) <- entries]
   f (UnkP _) = error "Cannot build unsolved type"