+++ ChangeLog.md
@@ -22,14 +22,16 @@ First hackage release
 
  - [ ] Print general types after subcommands
  - [ ] Allow spaces and no unecessary escape in input JSON data
+ - [ ] Remove extra space printed at the end of R JSON
  - [ ] Remove semicolon requirement
  - [ ] Haddock documentation
  - [ ] Update README
 
 0.32.0 [2020.10.xx]
 
- - [ ] Add record/table field access
+ - [.] Add record/table field access
  - [ ] Add record/table arithmetic sugar
+ - [x] Resolve most GHC warnings
 
 0.31.0 [2020.10.29]